backend/requirements.txt
aerich[toml]==0.8.1
configuretron==0.3.0
fastapi==0.115.0
google-auth==2.38.0
httpx==0.27.2
redis==5.2.1
tortoise-orm[asyncpg]==0.24.0
PyJWT==2.10.1
python-youtube==0.9.7
youtube_transcript_api==0.6.3
uvicorn==0.31.0